Adds Handlebars helpers
Throughout the course of working on the Halo: Mythic FoundryVTT
system, I found and wrote a number of helper functions that proved
essential for performing any sort of logic inside the Handlebars
templates themselves (conditionally rendering different sections,
iterating over a collection to render a section for each, etc). I
have introduced these here to save us as much grief as possible in
this project.